Project3/requirements.txt
absl-py==0.6.1
astor==0.7.1
certifi==2016.2.28
cycler==0.10.0
gast==0.2.0
graphviz==0.10.1
grpcio==1.17.0
h5py==2.8.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.3
numpy==1.15.4
Pillow==5.3.0
progressbar==2.5
protobuf==3.6.1
pydot==1.4.1
pydot-ng==2.0.0
pyparsing==2.3.1
PyQt5==5.11.3
PyQt5-sip==4.19.13
python-dateutil==2.8.0
PyYAML==5.1
scikit-learn==0.20.1
scipy==1.1.0
six==1.12.0
sklearn==0.0
tensorboard==1.10.0
tensorboardX==1.5
tensorflow-gpu==1.10.0
termcolor==1.1.0
torch==1.0.0
torchvision==0.2.1
Werkzeug==0.14.1
wincertstore==0.2
opencv-python==4.1.0.25
